rename: 패키지 명 변경 - client에서 aiModel로 변경
